Make dmenu.sh work on a Mac
I needed to tweak this to work on a Mac, because BSD `sed` doesn't
interpret `\n` as a newline. So this happens:

![screen shot 2017-03-04 at 8 30 35
am](https://cloud.githubusercontent.com/assets/305268/23580432/bf10377c-00b6-11e7-8c16-2bd8a06ea07c.png)

The easiest thing to do to fix it is to simply replace `sed` with `tr`.